refresh collection detail pages (bug 1103202/1103304)
