+++ src/dev.html
@@ -51,6 +51,7 @@
       <link rel="stylesheet" href="/media/css/feed/brands.styl.css">
       <link rel="stylesheet" href="/media/css/feed/collections.styl.css">
       <link rel="stylesheet" href="/media/css/feed/icon-grid.styl.css">
+      <link rel="stylesheet" href="/media/css/feed/landing.styl.css">
       <link rel="stylesheet" href="/media/css/feed/layout.styl.css">
       <link rel="stylesheet" href="/media/css/feed/mini-tiles.styl.css">
       <link rel="stylesheet" href="/media/css/feedback.styl.css">
