+++ src/media/css/feed/brands.styl
@@ -44,15 +44,6 @@
     bottom: 15px;
 }
 
-.brand-landing {
-    &.full > div {
-        padding-top: 10px;
-    }
-    .app-list {
-        margin-top: 0;
-    }
-}
-
 .brand-landing-icon {
     border-radius: 128px;
     height: 64px;
@@ -90,7 +81,7 @@
     z-index: -1;
 }
 
-@media $base-desktop {
+@media $base-tablet {
     .brand-landing-header-mobile {
         display: none;
     }
