+++ src/media/css/feed/landing.styl
@@ -1,0 +1,85 @@
+/*
+    Stylesheet for common styles for feed landing pages.
+    For example, the header type and app list padding.
+*/
+@import '../lib';
+
+.feed-landing-header {
+    display: block;
+}
+
+.feed-landing-header-inner {
+    max-width: 100%;
+    padding: 0 10px;
+}
+
+.feed-landing-bg-image {
+    margin: 0 auto;
+
+    // Change some stylings when a background image is present.
+    + .feed-landing-header-inner {
+        margin-top: 20px;
+
+        h1 {
+            margin-top: 0;
+        }
+    }
+}
+
+.feed-landing {
+    margin-top: 0;
+
+    &.full > div {
+        padding-top: 10px;
+    }
+    .app-list {
+        margin-top: 0;
+    }
+}
+
+.feed-landing-description {
+    text-align: center;
+}
+
+@media $base-tablet {
+    .feed-landing-header {
+        display-flex();
+        flex-wrap: nowrap;
+        justify-content: center;
+
+        h1 {
+            desktop_mega();
+            margin: 20px 0;
+            text-align: center;
+        }
+    }
+    .feed-landing-bg-image {
+        margin: 0 40px 0 0;
+
+        + .feed-landing-header-inner {
+            margin-top: 60px;
+
+            h1,
+            p {
+                text-align: left;
+            }
+        }
+    }
+    .feed-landing {
+        margin-top: 40px;
+    }
+    .brand-landing-header-desktop,
+    .collection-landing-header-desktop {
+        margin-top: 40px;
+    }
+}
+
+@media $base-desktop {
+    .feed-landing-header {
+        flex-wrap: wrap;
+    }
+    .brand-landing-header-desktop,
+    .collection-landing-header-desktop {
+        margin-top: 22px;  // To get 40px. Combine with 18px #page padding.
+    }
+}
