+++ src/media/css/feed/layout.styl
@@ -23,7 +23,6 @@
 @media screen and (max-width: 639px) {
     // One column.
     .home-feed,
-    .feed-landing.collection-landing,
     .feed-landing.shelf-landing {
         width: $feed-column-width;
     }
@@ -39,7 +38,6 @@
 
 @media screen and (min-width: 640px) {
     // Two columns at the most.
-    .feed-landing.collection-landing,
     .feed-landing.shelf-landing {
         width: $feed-column-width * 2;
     }
@@ -84,7 +82,6 @@
 }
 
 @media screen and (min-width: 640px) and (max-width: 709px) {
-    .feed-landing.collection-landing,
     .feed-landing.shelf-landing {
         padding-top: 18px;
     }
