+++ src/media/js/helpers_local.js
@@ -56,6 +56,12 @@ define('helpers_local',
         return text;
     };
 
+    filters.shadeHex = function(color, percent) {
+        // http://stackoverflow.com/questions/5560248/programmatically-lighten-or-darken-a-hex-color-or-rgb-and-blend-colors
+        var num = parseInt(color.slice(1),16), amt = Math.round(2.55 * percent), R = (num >> 16) + amt, G = (num >> 8 & 0x00FF) + amt, B = (num & 0x0000FF) + amt;
+        return "#" + (0x1000000 + (R<255?R<1?0:R:255)*0x10000 + (G<255?G<1?0:G:255)*0x100 + (B<255?B<1?0:B:255)).toString(16).slice(1);
+    };
+
     function has_installed(manifestURL) {
         return z.apps.indexOf(manifestURL) !== -1;
     }
